started working on a GradientMixin class
